Modified the getopt_port code to avoid MSVC warning and changed the version on the PyBind11 codes. Also hidden the GCC pragmas from non-gcc compilers to avoid other MSVC warnings.

+++ proshade/extern/getopt_port/getopt_port.c
@@ -189,7 +189,7 @@ int getopt_long_port ( int argc, char* const argv[], const char* optstring,
     /* If longindex is not NULL, it points to a variable which is set to the
        index of the long option relative to longopts. */
     if (longindex)
-      *longindex = (match - longopts);
+      *longindex = (int)( match - longopts );
 
     /* If flag is NULL, then getopt_long() shall return val.
        Otherwise, getopt_long() returns 0, and flag shall point to a variable
